# Conflicts:
#	stock-trading-server/target/classes/edu/rmit/sef/stocktradingserver/order/service/OrderHandler.class

+++ command-lib/src/main/java/edu/rmit/command/core/ICommandStore.java
@@ -4,8 +4,9 @@ package edu.rmit.command.core;
 import java.util.concurrent.CompletableFuture;
 
 public interface ICommandStore {
-    long getAsyncTaskCount();
-    long getTasKCount();
+    int getAsyncTaskCount();
+
+    int getTasKCount();
 
     <R, T extends ICommand<R>> CompletableFuture<R> execute(ICommandExecutionContext<T> context, ExecutionOptions options);
 }
